refactor: profileなどの直下にAnimationsやDimensionsの固定値をまとめたファイルを置いていたが、新たにprofileなどの直下にthemeというパッケージを作り、そこに置くように変更
