+++ composeApp/src/commonMain/kotlin/org/example/project/ui/feature/profile/theme/ProfileAnimations.kt
@@ -1,6 +1,6 @@
 @file:Suppress("MagicNumber")
 
-package org.example.project.ui.feature.profile
+package org.example.project.ui.feature.profile.theme
 
 internal data object ProfileAnimations {
     const val RatedDuration = 3000
