+++ composeApp/src/commonMain/kotlin/org/example/project/ui/feature/splash/theme/SplashAnimations.kt
@@ -1,6 +1,6 @@
 @file:Suppress("MagicNumber")
 
-package org.example.project.ui.feature.splash
+package org.example.project.ui.feature.splash.theme
 
 internal data object SplashAnimations {
     const val InitialProfileIconXOffset = -200f
