+++ composeApp/src/commonMain/kotlin/org/example/project/ui/feature/splash/theme/SplashDimensions.kt
@@ -1,5 +1,5 @@
 @file:Suppress("MagicNumber")
 
-package org.example.project.ui.feature.splash
+package org.example.project.ui.feature.splash.theme
 
 internal data object SplashDimensions
